feat: update lock extra fields

+++ src/bot/core.rs
@@ -2302,7 +2302,7 @@ rid|{}\nplatformID|0,1,1\ndeviceVersion|0\ncountry|jp\nhash|{}\nmac|{}\nwk|{}\nz
         for tile in &world.tile_map.tiles {
             if LOCK_ITEM_IDS.contains(&tile.fg_item_id) {
                 if let TileType::Lock { access_uids, .. } = &tile.tile_type {
-                    if access_uids.contains(&bot_uid) {
+                    if access_uids.contains(&(bot_uid as i32)) {
                         return true;
                     }
                 }
@@ -3017,4 +3017,3 @@ rid|{}\nplatformID|0,1,1\ndeviceVersion|0\ncountry|jp\nhash|{}\nmac|{}\nwk|{}\nz
         }
     }
 }
-
